Create a potential cron job file
This is an initial attempt to create a file to support cron jobs importing records.  The criteria at minimum is the filename.

The file was generated using the ZC file admin/currency_cron.php as a template and incorporating the initializing data of the main easypopulate_4.php file.  Doing this has indicated that some modularity should definitely be incorporated into that file.